각각 다른 형태로 구현해 보았던 코드를 모두 삭제 후 최신것으로 갱신
